+++ packages/dart_pptx/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- Fixing hidden files not uploading to pub.dev
+
 ## 0.1.0
 
 - Added slide templates from keynote