+++ packages/flutter_pptx/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.1
+
+- Fixing hidden files not uploading to pub.dev
+- Updating `dart_pptx` dependency to `0.1.1`
+
 ## 0.1.0
 
 - Adding support for widget slides